chore: Move UIElement debugging methods to separate partial
